Backend_api/requirements.txt
flask
flask_sqlalchemy
pymongo
flask_cors
flask_pymongo
PyJWT
requests
matplotlib
